+++ .cirrus.yml
@@ -9,8 +9,8 @@ task:
     . $HOME/.cargo/env
   cargo_cache:
     folder: $CARGO_HOME/registry
-  build_script: env PATH="$HOME/.cargo/bin:$PATH" cargo build
-  test_script: env PATH="$HOME/.cargo/bin:$PATH" cargo test
+  build_script: env PATH="$HOME/.cargo/bin:$PATH" cargo build --all && cargo build --examples
+  test_script: env PATH="$HOME/.cargo/bin:$PATH" cargo test --all
   before_cache_script: rm -rf $CARGO_HOME/registry/index
 
 
@@ -24,6 +24,20 @@ task:
     . $HOME/.cargo/env
   cargo_cache:
     folder: $CARGO_HOME/registry
-  build_script: env PATH="$HOME/.cargo/bin:$PATH" cargo build
-  test_script: env PATH="$HOME/.cargo/bin:$PATH" cargo test
+  build_script: env PATH="$HOME/.cargo/bin:$PATH" cargo build --all && cargo build --examples
+  test_script: env PATH="$HOME/.cargo/bin:$PATH" cargo test --all
+  before_cache_script: rm -rf $CARGO_HOME/registry/index
+
+task:
+  matrix:
+    - container:
+        image: rust:latest
+    - allow_failures: true
+      container:
+        image: rustlang/rust:nightly
+  cargo_cache:
+    folder: $CARGO_HOME/registry
+    fingerprint_script: cat Cargo.lock
+  build_script: cargo build --all && cargo build --examples
+  test_script: cargo test --all
   before_cache_script: rm -rf $CARGO_HOME/registry/index
