+++ Cargo.toml
@@ -5,16 +5,22 @@ authors = [
    "Johannes Lundberg <johalun0@gmail.com>",
    "Ivan Temchenko <ivan.temchenko@yandex.ua>",
    "Fabian Freyer <fabian.freyer@physik.tu-berlin.de>"
-   ]
+]
 description = "Simplified interface to libc::sysctl"
 keywords = ["sysctl", "freebsd", "macos"]
 license = "MIT"
 readme = "README.md"
 repository = "https://github.com/johalun/sysctl-rs"
 documentation = "https://johalun.github.io/sysctl-rs/index.html"
+edition = "2018"
 
 [dependencies]
-libc = "^0.2.34"
+failure   = "^0.1"
+
+[target.'cfg(any(target_os = "macos", target_os = "freebsd"))'.dependencies]
+libc      = "^0.2.34"
 byteorder = "^1.0.0"
-failure = "^0.1.1"
-bitflags = "^1.0"
+bitflags  = "^1.0"
+
+[features]
+default = [ ]
