+++ examples/struct.rs
@@ -1,23 +1,20 @@
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 extern crate libc;
 extern crate sysctl;
-
-#[cfg(not(target_os = "macos"))]
-use libc::c_int;
-#[cfg(not(target_os = "macos"))]
-use std::mem;
 
 // Copied from /usr/include/sys/time.h
 #[derive(Debug)]
 #[repr(C)]
-#[cfg(not(target_os = "macos"))]
+#[cfg(target_os = "freebsd")]
 struct ClockInfo {
-    hz: c_int,   /* clock frequency */
-    tick: c_int, /* micro-seconds per hz tick */
-    spare: c_int,
-    stathz: c_int, /* statistics clock frequency */
-    profhz: c_int, /* profiling clock frequency */
+    hz: libc::c_int,   /* clock frequency */
+    tick: libc::c_int, /* micro-seconds per hz tick */
+    spare: libc::c_int,
+    stathz: libc::c_int, /* statistics clock frequency */
+    profhz: libc::c_int, /* profiling clock frequency */
 }
-#[cfg(not(target_os = "macos"))]
+
+#[cfg(target_os = "freebsd")]
 fn main() {
     let ctl = sysctl::Ctl::new("kern.clockrate").expect("could not get sysctl: kern.clockrate");
 
@@ -32,7 +29,7 @@ fn main() {
 
     if let sysctl::CtlValue::Struct(val) = val_enum {
         // Make sure we got correct data size
-        assert_eq!(mem::size_of::<ClockInfo>(), val.len());
+        assert_eq!(std::mem::size_of::<ClockInfo>(), val.len());
         let val_ptr: *const u8 = val.as_ptr();
         let struct_ptr: *const ClockInfo = val_ptr as *const ClockInfo;
         let struct_ref: &ClockInfo = unsafe { &*struct_ptr };
@@ -40,5 +37,5 @@ fn main() {
     }
 }
 
-#[cfg(target_os = "macos")]
+#[cfg(not(target_os = "freebsd"))]
 fn main() {}
