+++ examples/value.rs
@@ -1,42 +1,29 @@
 extern crate sysctl;
 
-use sysctl::{Ctl, CtlValue};
+use std::io;
 
-#[cfg(not(target_os = "macos"))]
 fn main() {
-    let ctl = Ctl::new("kern.osrevision").expect("could not get sysctl");
+    #[cfg(any(target_os = "macos", target_os = "freebsd"))]
+    const KEY: &str = "net.inet.ip.forwarding";
+    #[cfg(any(target_os = "android", target_os = "linux"))]
+    const KEY: &str = "/proc/sys/net/ipv4/ip_forward";
 
-    let name = ctl.name().expect("could not get name");
+    #[cfg(any(target_os = "macos", target_os = "freebsd"))]
+    let ret = sysctl::value(KEY)
+        .map(|v| v == sysctl::CtlValue::Int(1))
+        .map_err(|_| io::Error::last_os_error());
 
-    println!("\nRead sysctl {}", name);
+    #[cfg(any(target_os = "android", target_os = "linux"))]
+    let ret = sysctl::value(KEY)
+        .map(|v| v == sysctl::CtlValue::String("1\n".to_string()))
+        .map_err(|_| io::Error::last_os_error());
 
-    let description = ctl.description().expect("could not get description");
-
-    println!("Description: {:?}", description);
-
-    let val_enum = ctl.value().expect("could not get sysctl value");
-
-    if let CtlValue::Int(val) = val_enum {
-        println!("Value: {}", val);
-    }
-}
-
-#[cfg(target_os = "macos")]
-fn main() {
-    // on macos the `name` and `newp` parameters of the sysctl(3)
-    // syscall API are not marked `const`. This means the sysctl
-    // structure has to be mutable.
-    let ctl = Ctl::new("kern.osrevision").expect("could not get sysctl");
-
-    let name = ctl.name().expect("could not get name");
-
-    println!("\nRead sysctl {}", name);
-
-    // sysctl descriptions are not available on macos.
-
-    let val_enum = ctl.value().expect("could not get sysctl value");
-
-    if let CtlValue::Int(val) = val_enum {
-        println!("Value: {}", val);
+    match ret {
+        Ok(val) => {
+            println!("The `{}` value is: {}", KEY, val);
+        }
+        Err(e) => {
+            println!("ERROR: {:?}", e);
+        }
     }
 }
