+++ examples/value_as.rs
@@ -1,26 +1,31 @@
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 extern crate libc;
 extern crate sysctl;
 
-use libc::c_int;
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 use std::fmt;
 
 // Copied from /usr/include/sys/time.h
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 #[derive(Debug)]
 #[repr(C)]
 struct ClockInfo {
-    hz: c_int,   /* clock frequency */
-    tick: c_int, /* micro-seconds per hz tick */
-    spare: c_int,
-    stathz: c_int, /* statistics clock frequency */
-    profhz: c_int, /* profiling clock frequency */
+    hz: libc::c_int,   /* clock frequency */
+    tick: libc::c_int, /* micro-seconds per hz tick */
+    spare: libc::c_int,
+    stathz: libc::c_int, /* statistics clock frequency */
+    profhz: libc::c_int, /* profiling clock frequency */
 }
 
 // Copied from /usr/include/sys/resource.h
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 #[repr(C)]
 struct LoadAvg {
     ldavg: [u32; 3],
     fscale: u64,
 }
+
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 impl fmt::Debug for LoadAvg {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         let s = self.fscale as f32;
@@ -34,6 +39,7 @@ impl fmt::Debug for LoadAvg {
     }
 }
 
+#[cfg(any(target_os = "macos", target_os = "freebsd"))]
 fn main() {
     // Generic type to pass to function will be inferred if not specified on RHS
     println!("\nRead sysctl kern.clockrate as struct directly");
@@ -51,3 +57,6 @@ fn main() {
         .expect("could not read sysctl as LoadAvg");
     println!("{:?}", val);
 }
+
+#[cfg(not(any(target_os = "macos", target_os = "freebsd")))]
+fn main() {}
